Completed first section of Jobbatical
This is a working copy of my part for Jobbatical. It is responsive and when you minimise it, a button pops out which when clicked slides in a side bar from the right. A second click closes the side bar. Aesthetically some font should be changed, button be made smaller etc. But I decided to upload this so we can move on and finish asap. Our main focus is on functionality.